test compute schedule
the compute schedule function works now I think. I modified a couple of the set color functions to be able to take a color, leaving the previous colors as defaults. I added compute_MST_nx which is a version that returns an nx graph instead of a manim graph. I also have a test for creating a tile scene with the stacked layers/colors, (and it needs to connect some edges graphically I think, but that fixed some errorthough I don't know if it will look correct, so that may need some modification. There are index errors with the size of the tile, eg over 25. I tamped it down to a smaller support set, but that didn't solve it.